# backend/nosd/openapi.yaml
openapi: 3.0.3
info:
  title: NithronOS API
  version: 0.1.0
  description: API for nOS backend
servers:
  - url: http://127.0.0.1:9000
paths:
...
  /api/pools/{id}/snapshots:
    get:
      summary: List snapshots
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snapshot'
    post:
      summary: Create snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subvol: { type: string }
                name: { type: string }
              required: [subvol]
      responses:
        '200':
          description: OK
...
  /api/updates/rollback:
    post:
      summary: Roll back a prior updates transaction using its snapshots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tx_id: { type: string }
                confirm: { type: string }
              required: [tx_id, confirm]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Snapshot:
      type: object
      properties:
        path: { type: string }
        name: { type: string }
        readonly: { type: boolean }
...